avoid redundant memory accesses
The percent-encoding of a byte is 3 bytes long, and if the output string
ends right after it, a NUL is also appended, so a 4-byte append to the
output is always valid. Therefore the percent-encoding can be treated as
a single 32-bit value which can not only be read and written in a single
instruction but also zero-tested as a single in-register value.